feat: course API 모델 분리
